#36 update models
