--- src/models/IPatternUnit.ts
@@ -1,6 +1,0 @@
-export interface IPatternUnit {
-  startDelay: number;
-  duration: number;
-  weakMagnitude: number;
-  strongMagnitude: number;
-}
