--- src/models/TPatternUnit.ts
@@ -1,6 +1,0 @@
-export type TPatternUnit = {
-  startDelay: number;
-  duration: number;
-  weakMagnitude: number;
-  strongMagnitude: number;
-};
