+++ src/store/models/index.ts
@@ -1,0 +1,3 @@
+export * from './GamepadsState';
+export * from './PatternState'
+export * from './RootState'
